refactor: use glslang compiler singleton pattern
The glslang compiler implementation was previous reworked to avoid using
a singleton pattern in its API. Upon further inspection, this design is
problematic because it obfuscates creation of the glslang process. To
fix this, the singleton instance pattern is being reintroduced which
should make this behavior more explicit when using this class.

Less important is that this refactor also allows for a consistent naming
convention where classes are prefixed by type rather than using a
secondary namespace (e.g., KtxTexture and likely in the future
GltfScene).