feat: add texture abstraction
`gltf_scene.cppm` is an absolute unit of a file that contains many glTF
abstractions that should be refactored into self-contained units so that
they can be more easily managed and extended. This commit makes a first
attempt at this by creating a dedicated texture abstraction responsible
for loading an image file from host memory and copying it to device
memory as a Vulkan image with VK_FORMAT_FEATURE_SAMPLED_IMAGE_BIT
support. To avoid potential confusion, this has been merged with the
existing `ktx_texture` module which is an implementation detail that
does not warrant an exposed API surface.